Update avatar & preview
